refactor: customer페이지에서 사용되는 함수및 값들 해당 페이지로 이동
